Fixing the DHCP Server.....
I was so happy it worked correctly. But it just didn't and I didn't know. What happened?
I setup the Pihole DHCP thinking it would work. Rebooted the Raspberry Pi, and everything still worked.
But it wasn't: I set it up that the ip adresses should be refreshing every hour. In the first hour after changing, everything worked since the ip adresses were still valid.
but after that hour, they weren't. suddenly no access to anything. Took a whole hour to find and fix the problem.